refactor : separate sections
