+++ src/main/java/nextstep/subway/line/domain/Line.java
@@ -1,20 +1,19 @@
 package nextstep.subway.line.domain;
 
-import java.util.ArrayList;
-import java.util.LinkedList;
 import java.util.List;
-import java.util.Map;
-import java.util.stream.Collectors;
 
+import javax.persistence.CascadeType;
 import javax.persistence.Column;
 import javax.persistence.Entity;
 import javax.persistence.GeneratedValue;
 import javax.persistence.GenerationType;
 import javax.persistence.Id;
 import javax.persistence.OneToMany;
+import javax.persistence.Transient;
 
 import nextstep.subway.common.BaseEntity;
 import nextstep.subway.section.domain.Section;
+import nextstep.subway.section.domain.Sections;
 import nextstep.subway.station.domain.Station;
 
 @Entity
@@ -29,8 +28,11 @@ public class Line extends BaseEntity {
 	@Column
 	private String color;
 
-	@OneToMany(mappedBy = "line")
-	private List<Section> sections = new LinkedList<>();
+	@OneToMany(mappedBy = "line", cascade = CascadeType.ALL, orphanRemoval = true)
+	private List<Section> sectionList = new Sections().getSections();
+
+	@Transient
+	private Sections sections = new Sections();
 
 	protected Line() {
 	}
@@ -57,49 +59,12 @@ public class Line extends BaseEntity {
 		return color;
 	}
 
-	public void addSections(Section section) {
+	public void addSection(Section section) {
 		sections.add(section);
 		section.setLine(this);
 	}
 
 	public List<Station> getOrderedStations() {
-		Map<Station, Station> order = getStationOrder();
-		return convertToOrderedList(order);
-	}
-
-	private Map<Station, Station> getStationOrder() {
-		return sections.stream().collect(Collectors.toMap(Section::getUpStation, Section::getDownStation));
-	}
-
-	private List<Station> convertToOrderedList(Map<Station, Station> order) {
-		List<Station> stations = new ArrayList<>();
-		Station key = getStartStation(order);
-
-		stations.add(key);
-
-		while (order.containsKey(key)) {
-			stations.add(order.get(key));
-			key = order.get(key);
-		}
-
-		return stations;
-	}
-
-	private Station getStartStation(Map<Station, Station> order) {
-		Station key = null;
-
-		for (Station station : order.keySet()) {
-			if (!order.containsValue(station)) {
-				key = station;
-			}
-		}
-		return key;
-	}
-
-	public Long delete() {
-		sections.stream().forEach(section -> section.setLine(null));
-		sections = null;
-
-		return id;
+		return sections.getOrderedStations();
 	}
 }
